Changes to make the library easily relocatable
If someone wants to use playbox2d in their own project, this makes it as easy as a simple `include playbox2d/playbox2d.mk` in their makefile and calling `registerPlaybox()` in an init event.

Include statements are now fully qualified to avoid clashes with other libraries (with something like platform.h for example).

# Conflicts:
#	Makefile